feat: función para registrar archivos ya archivarlos en la carpeta /uploads

+++ src/main/java/com/negocio/ProyectoHerramientaDeDesarrolloWeb/controller/CursoController.java
@@ -33,7 +33,7 @@ public class CursoController {
     ///
     /// ////////////////////////////////
     @GetMapping("/obtener/{id}")
-    public ResponseEntity<?> obtenerPorId(@PathVariable int id){
+    public ResponseEntity<?> obtenerPorId(@PathVariable Long id){
         return cursoService.obtenerPorId(id);
     }
     @GetMapping("/obtener")
@@ -48,7 +48,7 @@ public class CursoController {
     /// ////////////////////////////////
 
     @PutMapping("/eliminar/{id}")
-    public ResponseEntity<?> obtenerTodos(@PathVariable int id){
+    public ResponseEntity<?> obtenerTodos(@PathVariable Long id){
         return cursoService.eliminarPorId(id);
     }
 }
